+++ backend/src/scalar_types.rs
@@ -19,6 +19,8 @@ pub type ModuleReference = String;
 pub type TransactionIndex = i64;
 pub type AccountIndex = i64;
 pub type MetadataUrl = String;
+pub type PltIndex = i64;
+pub type TokenIndex = i64;
 
 /// The UnsignedLong scalar type represents a unsigned 64-bit numeric
 /// non-fractional value greater than or equal to 0.